use size_t for raw_getrandom

+++ rng/unix/mc_getrandom_stubs.c
@@ -21,7 +21,7 @@
 # define getrandom(buf, len, flags) getrandom((buf), (len), (flags))
 # endif
 
-void raw_getrandom (uint8_t *data, uint32_t len) {
+void raw_getrandom (uint8_t *data, size_t len) {
   size_t off = 0;
   ssize_t r = 0;
   while (off < len) {
@@ -41,9 +41,9 @@ void raw_getrandom (uint8_t *data, uint32_t len) {
 #endif
 #include <sys/param.h>
 
-void raw_getrandom (uint8_t *data, uint32_t len) {
+void raw_getrandom (uint8_t *data, size_t len) {
   size_t rlen = 0;
-  for (uint32_t i = 0; i <= len; i += 256) {
+  for (size_t i = 0; i <= len; i += 256) {
     rlen = MIN(256, len - i);
     if (getentropy(data + i, rlen) == -1) uerror("getentropy", Nothing);
   }
@@ -63,7 +63,7 @@ void raw_getrandom (uint8_t *data, uint32_t len) {
 #include <ntstatus.h>
 #include <bcrypt.h>
 
-void raw_getrandom(uint8_t *data, uint32_t len) {
+void raw_getrandom(uint8_t *data, size_t len) {
    NTSTATUS Status;
    Status = BCryptGenRandom(NULL, data, len, BCRYPT_USE_SYSTEM_PREFERRED_RNG);
    if (Status != STATUS_SUCCESS)
